@@ -6,9 +6,10 @@ import 'package:flutter/material.dart';
 
 class CenterLineDivider extends PositionComponent
     with HasGameReference<PongGame> {
-  CenterLineDivider(
-      {required this.dividerColor, required this.isDividerContinuous})
-      : super();
+  CenterLineDivider({
+    required this.dividerColor,
+    required this.isDividerContinuous,
+  }) : super();
   final Color dividerColor;
   final bool isDividerContinuous;
 
@@ -16,18 +17,22 @@ class CenterLineDivider extends PositionComponent
   FutureOr<void> onLoad() async {
     super.onLoad();
     if (isDividerContinuous) {
-      await add(RectangleComponent(
-        position: Vector2((game.width / 2) - 2, 0),
-        size: Vector2(5, game.height),
-        paint: dividerColor.filledPaint(),
-      ));
+      await add(
+        RectangleComponent(
+          position: Vector2((game.width / 2) - 2, 0),
+          size: Vector2(5, game.height),
+          paint: dividerColor.filledPaint(),
+        ),
+      );
     } else {
       await addAll(
         List.generate(
           30,
           (index) => RectangleComponent(
-            position: Vector2((game.width / 2) - 2,
-                (game.height / 30) * index + (index * 10)),
+            position: Vector2(
+              (game.width / 2) - 2,
+              (game.height / 30) * index + (index * 10),
+            ),
             size: Vector2(5, game.height / 30),
             paint: dividerColor.filledPaint(),
           ),
