@@ -1,6 +1,9 @@
+import 'dart:io';
+
 import 'package:easy_pong/functions.dart';
 import 'package:easy_pong/widgets/tile_button.dart';
 import 'package:flame/flame.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -36,13 +39,27 @@ class HomeScreen extends StatelessWidget {
                     await Flame.device.setPortrait();
                   },
                 ),
+                if (!kIsWeb && Platform.isAndroid) ...[
+                  const SizedBox(height: 20),
+                  TileButton(
+                    titleText: "LAN Multiplayer",
+                    width: isPhone() ? 250 : 350,
+                    onTap: () async {
+                      await Navigator.of(
+                        context,
+                      ).pushNamed('/real_time_multiplayer');
+                      await Flame.device.setPortrait();
+                    },
+                  ),
+                ],
                 const SizedBox(height: 20),
                 TileButton(
                   titleText: "Play vs Computer",
                   width: isPhone() ? 250 : 350,
                   onTap: () async {
-                    await Navigator.of(context)
-                        .pushNamed('/computer_difficulty');
+                    await Navigator.of(
+                      context,
+                    ).pushNamed('/computer_difficulty');
                     await Flame.device.setPortrait();
                   },
                 ),
